requirements.txt
Pillow>=9.5.0
numpy>=1.21.0